requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.2.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kucoin-python==1.0.10
Mako==1.1.6
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.22.0
pandas==1.3.5
pandas-ta==0.3.14b0
plotly==5.5.0
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
tenacity==8.0.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.7
websocket==0.2.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
